fix: fixed issue with behaviour in case of incorrect ID

+++ backend/rust/order_service/src/repository.rs
@@ -221,13 +221,17 @@ impl OrderRepository {
     /// - `Ok`         - in case of success.
     /// - `SQLx error` - otherwise.
     pub async fn delete(&self, id: i64) -> Result<(), sqlx::Error> {
-        let query = r#"DELETE FROM Orders WHERE order_id = ?"#;
-        sqlx::query(query).bind(&id).execute(&self.pool).await?;
+        let query  = r#"DELETE FROM Orders WHERE order_id = ?"#;
+        let result = sqlx::query(query).bind(&id).execute(&self.pool).await?;
+
+        // Handle incorrect order ID.
+        if result.rows_affected() == 0 {
+            return Err(sqlx::Error::RowNotFound);
+        }
 
         let query = r#"DELETE FROM OrderItems WHERE order_id = ?"#;
         sqlx::query(query).bind(&id).execute(&self.pool).await?;
 
-        println!("[order-service] Deleted order with ID: {}", id);
         Ok(())
     }
